% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{draw_rownames}
\alias{draw_rownames}
\title{draw_rownames}
\usage{
draw_rownames(.data)
}
\arguments{
\item{.data}{dataframe with rownames}
}
\value{
tibble
}
\description{
to make dataframes more tibble compatible by saving rownames to column
}
