% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%T\%}
\alias{\%T\%}
\title{infix T Pipe}
\usage{
c(1,2,3) %T% list(first=partial(head, n=1), last=partial(tail, n=1))
}
\arguments{
\item{d}{data to pass into funcs}

\item{funcs}{named list of functions that take data in same form of d}
}
\value{
list of results, preserving names
}
\description{
Takes a list of functions to apply to an input
}
