% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%T>\%}
\alias{\%T>\%}
\title{infix lapply}
\usage{
c(1,2,3) %T>% function(x) {return(x ** 2)}
}
\arguments{
\item{l}{list of inputs}

\item{f}{function to apply accross list}
}
\value{
list of return values
}
\description{
infix lapply
}
