% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{infixcompose}
\alias{infixcompose}
\alias{\%|\%}
\title{function composition}
\usage{
f \%|\% g
}
\arguments{
\item{f}{function with one input parameter}

\item{g}{function with one input parameter}
}
\value{
function
}
\description{
compose functions with single parameter
}
