% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{partial}
\alias{partial}
\title{partial evaluation}
\usage{
partial(f, ...)
}
\arguments{
\item{f}{function}

\item{...}{loaded partial parameters after first param}
}
\value{
g <- function(X) f(X, ...)
}
\description{
partial evaluation of function, consistant with dplyr first param
}
